ice_breaker_langchain/react_langchain/requirements.txt
langchain
openai
black
python-dotenv
